+++ src/components/FiscalYear/FiscalYearView.js
@@ -2,15 +2,8 @@ import React, { Component } from 'react';
 import _ from 'lodash';
 import PropTypes from 'prop-types';
 import queryString from 'query-string';
-// Components and Pages
-import Layer from '@folio/stripes-components/lib/Layer';
-import Pane from '@folio/stripes-components/lib/Pane';
-import PaneMenu from '@folio/stripes-components/lib/PaneMenu';
-import { Row, Col } from '@folio/stripes-components/lib/LayoutGrid';
-import Icon from '@folio/stripes-components/lib/Icon';
-import IconButton from '@folio/stripes-components/lib/IconButton';
-import IfPermission from '@folio/stripes-components/lib/IfPermission';
-import KeyValue from '@folio/stripes-components/lib/KeyValue';
+import { Layer, Pane, PaneMenu, Icon, IconButton, IfPermission, KeyValue, Row, Col } from '@folio/stripes-components';
+import { withTags } from '@folio/stripes-smart-components/lib/Tags';
 import FiscalYearPane from './FiscalYearPane';
 import ConnectionListing from '../ConnectionListing';
 import FormatDate from '../../Utils/FormatDate';
@@ -33,7 +26,10 @@ class FiscalYearView extends Component {
     paneWidth: PropTypes.oneOfType([
       PropTypes.string,
       PropTypes.number
-    ])
+    ]),
+    notesToggle: PropTypes.func,
+    tagsToggle: PropTypes.func,
+    tagsEnabled: PropTypes.bool
   }
 
   static getDerivedStateFromProps(nextProps, prevState) {
@@ -108,11 +104,29 @@ class FiscalYearView extends Component {
   }
 
   render() {
-    const { location } = this.props;
+    const { location, tagsEnabled } = this.props;
     const initialValues = this.getData();
     const query = location.search ? queryString.parse(location.search) : {};
+    const tags = ((initialValues && initialValues.tags) || {}).tagList || [];
     const detailMenu = (
       <PaneMenu>
+        {
+          tagsEnabled &&
+          <IconButton
+            icon="tag"
+            title="showTags"
+            id="clickable-show-tags"
+            onClick={this.props.tagsToggle}
+            badgeCount={tags.length}
+            aria-label="showTags"
+          />
+        }
+        <IconButton
+          icon="comment"
+          id="clickable-show-notes"
+          onClick={this.props.notesToggle}
+          aria-label="showNotes"
+        />
         <IfPermission perm="fiscal_year.item.put">
           <IconButton
             icon="edit"
@@ -142,19 +156,19 @@ class FiscalYearView extends Component {
       <Pane id="pane-fiscalyeardetails" defaultWidth={this.props.paneWidth} paneTitle={_.get(initialValues, ['name'], '')} lastMenu={detailMenu} dismissible onClose={this.props.onClose}>
         <Row>
           <Col xs={4}>
-            <KeyValue label="name" value={_.get(initialValues, ['name'], '')} />
+            <KeyValue label="Name" value={_.get(initialValues, ['name'], '')} />
           </Col>
           <Col xs={4}>
-            <KeyValue label="code" value={_.toString(_.get(initialValues, ['code'], ''))} />
+            <KeyValue label="Code" value={_.toString(_.get(initialValues, ['code'], ''))} />
           </Col>
           <Col xs={4}>
             <KeyValue label="Description" value={_.get(initialValues, ['description'], '')} />
           </Col>
           <Col xs={4}>
-            <KeyValue label="period start" value={startDate} />
+            <KeyValue label="Period Begin Date" value={startDate} />
           </Col>
           <Col xs={4}>
-            <KeyValue label="period end" value={endDate} />
+            <KeyValue label="Period End Date" value={endDate} />
           </Col>
           {
             isLedgerData &&
@@ -200,4 +214,4 @@ class FiscalYearView extends Component {
   }
 }
 
-export default FiscalYearView;
+export default withTags(FiscalYearView);