+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change history for ui-finance
 
+## 1.1.1
+* Implement tags into Ledger, Funds, Budget, Fiscal Year and. ref: UIF-17, UIF-18, UIF-19 and UIF-20.
+* Added asterisk to required field in Budget. ref: UIF-13.
+* Added asterisk and required fiscal year start date and end date. ref: UIF-9.
+
 ## [1.0.1](https://github.com/folio-org/ui-vendors/tree/v1.0.1) (2018-09-17)
 * Update stripes-form version. Fixes UIV-8.